refactoring: use pytest for diffusers tests
