+++ .github/workflows/test_inf2_transformers.yml
@@ -35,10 +35,10 @@ jobs:
         uses: ./.github/actions/prepare_venv
       - name: Install optimum-neuron
         uses: ./.github/actions/install_optimum_neuron
-      - name: Run inference tests
+      - name: Run transformers inference tests
         run: |
           source aws_neuron_venv_pytorch/bin/activate
-          HF_TOKEN=${{ secrets.HF_TOKEN_OPTIMUM_NEURON_CI }} pytest -m is_inferentia_test tests/inference --forked -q
+          HF_TOKEN=${{ secrets.HF_TOKEN_OPTIMUM_NEURON_CI }} pytest -m is_inferentia_test tests/inference/transformers
       - name: Run pipelines tests
         run: |
           source aws_neuron_venv_pytorch/bin/activate
