+++ tests/inference/diffusers/test_flux.py
